test(fs): move 29184 tests into fs-oom.test.ts
Per CLAUDE.md, test/regression/issue/N.test.ts is reserved for true
regressions (worked, then broke). #29184 was never correct — the
lower-bound early-throw formula was wrong from inception — so the
tests belong in the module's existing file, test/js/node/fs/fs-oom.test.ts,
which already uses the identical memfd_create +
setSyntheticAllocationLimitForTesting pattern for the sync readFile path.

Adds a scoped describe block for the async readFile coverage (hex memfd
boundary, /dev/urandom issue repro, small-file sanity check) with a
beforeAll/afterAll that restores the synthetic allocation limit so the
describe doesn't leak state into later tests in the same process.
